Fix show alert before open twitter app
