+++ PyConJP2016/Resource/Localizable/Base.lproj/Localizable.strings
@@ -1,0 +1,15 @@
+/* 
+  Localizable.strings
+  PyConJP2016
+
+  Created by Yutaro Muta on 9/17/16.
+  Copyright © 2016 PyCon JP. All rights reserved.
+*/
+
+/**
+ AlertController
+ */
+
+"Cancel" = "Cancel";
+
+"Open" = "Open";
