+++ PyConJP2016/Resource/Localizable/ja.lproj/URLScheme.strings
@@ -1,0 +1,15 @@
+/* 
+  URLScheme.strings
+  PyConJP2016
+
+  Created by Yutaro Muta on 9/17/16.
+  Copyright © 2016 PyCon JP. All rights reserved.
+*/
+
+/**
+ TwitterURLSchemeType
+ */
+
+"twitterAlertTitle" = "Twiiterアプリを起動します";
+
+"twitterAlertMessage" = "%@ を開きます";
